+++ src/components/general/structure/Container/index.ts
@@ -1,0 +1,1 @@
+export * from './Container.component';
