+++ src/components/general/structure/Content/index.ts
@@ -1,0 +1,1 @@
+export * from './Content.component';
