+++ src/components/general/structure/Footer/index.ts
@@ -1,0 +1,1 @@
+export * from './Footer.component';
