+++ src/components/general/structure/ViewBase/index.ts
@@ -1,0 +1,1 @@
+export * from './ViewBase.component';
