+++ src/screens/groups/groups.screen.tsx
@@ -1,5 +1,5 @@
 import { SectionHeader } from '@moneyboy/components/general/lists/SectionHeader';
-import { ViewBase } from '@moneyboy/components/general/structure/viewBase';
+import { ViewBase } from '@moneyboy/components/general/structure/ViewBase';
 import { GroupListItem } from '@moneyboy/components/groups/groupListItem';
 import React from 'react';
 import { DefaultSectionT, SectionList, SectionListRenderItemInfo } from 'react-native';
