+++ src/screens/history/history.screen.tsx
@@ -1,7 +1,7 @@
 import { SectionHeader } from '@moneyboy/components/general/lists/SectionHeader';
 import { MoneyDiff, MoneyDiffProps } from '@moneyboy/components/general/payments/MoneyDiff';
-import { Content } from '@moneyboy/components/general/structure/content';
-import { ViewBase } from '@moneyboy/components/general/structure/viewBase';
+import { Content } from '@moneyboy/components/general/structure/Content';
+import { ViewBase } from '@moneyboy/components/general/structure/ViewBase';
 import { PescaContext } from '@moneyboy/contexts/pescaContext';
 import { useAuth } from '@moneyboy/hooks/useAuth';
 import { useStorage } from '@moneyboy/hooks/useStorage';
