+++ src/styles/content.ts
@@ -26,7 +26,7 @@ export const createContentStyles = (props?: StylingProps) => {
     dp24: mix('ffffff', base.toString(16), 16),
   };
 
-  // TODO lome: when introducing dark-mode, use several stages of content
+  // TODO lome: when introducing dark-mode, use several stages of Content
   return {
     background,
     separator,
