+++ tsconfig.json
@@ -40,7 +40,7 @@
     "paths": {
       "@moneyboy/*": ["src/*"],
     },
-    // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
+    // "rootDirs": [],                        /* List of root folders whose combined Content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
     // "types": [],                           /* Type declaration files to be included in compilation. */
     "allowSyntheticDefaultImports": true      /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */,
